Fixed fragmented code
